Struct wasmprinter::Config

pub struct Config { /* private fields */ }

Configuration used to print a WebAssembly binary.

This structure is used to control the overal structure of how wasm binaries are printed and tweaks various ways that configures the output.

Implementations

impl Config

pub fn new() -> Self

Creates a new Config object that’s ready to start printing wasm binaries to strings.

pub fn print_offsets(&mut self, print: bool)

Whether or not to print binary offsets of each item as comments in the text format whenever a newline is printed.

pub fn print_skeleton(&mut self, print: bool)

Whether or not to print only a “skeleton” which skips function bodies, data segment contents, element segment contents, etc.

pub fn name_unnamed(&mut self, enable: bool)

Assign names to all unnamed items.

If enabled then any previously unnamed item will have a name synthesized that looks like $#func10 for example. The leading # indicates that it’s wasmprinter-generated. The func is the namespace of the name and provides extra context about the item when referenced. The 10 is the local index of the item.

Note that if the resulting text output is converted back to binary the resulting name custom section will not be the same as before.

pub fn print(&self, wasm: &[u8], result: &mut impl Print) -> Result<()>

Prints a WebAssembly binary into a String

This function takes an entire wasm binary blob and will print it to the WebAssembly Text Format and return the result as a String.

pub fn offsets_and_lines<'a>( &self, wasm: &[u8], storage: &'a mut String, ) -> Result<impl Iterator<Item = (Option<usize>, &'a str)> + 'a>

Get the line-by-line WAT disassembly for the given Wasm, along with the binary offsets for each line.