Crate wasmprinter

A crate to convert a WebAssembly binary to its textual representation in the WebAssembly Text Format (WAT).

This crate is intended for developer toolchains and debugging, supporting human-readable versions of a wasm binary. This can also be useful when developing wasm toolchain support in Rust for various purposes like testing and debugging and such.

Structs

Config

Configuration used to print a WebAssembly binary.

PrintFmtWrite

An adapter between the std::fmt::Write trait and Print.

PrintIoWrite

An adapter between the std::io::Write trait and Print.

PrintTermcolor

An adapter between the std::fmt::Write trait and termcolor::WriteColor.

Traits

Print

Trait used to print WebAssembly modules in this crate.

Functions

print_bytes

Prints an in-memory wasm binary blob into an in-memory String which is its textual representation.

print_file

Reads a WebAssembly file from the filesystem and then prints it into an in-memory String.